[CMA-41] fix : SetClimerProfile 갤러리 접근 로직 IntroActivity 로 이동
